requirements.txt
Flask==3.0.2
python-pptx==0.6.23
Werkzeug==3.0.1
gunicorn==21.2.0
